Add focusin on price_format
I propose to add a focusin event on jquery.priceformat.js

This is simple, but useful when the project made a calculation and the result of calculation will be automatically put inside a textbox.

Example:
By adding the focusin event like what I proposed, the project just need to add a

document.getElementById("anyIDforResultTextBox").focus()

After that, the result of calculation will be following the needed priceformat on the textbox.